% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_backlog_history}
\alias{get_backlog_history}
\title{Get Historical Backlog Size}
\usage{
get_backlog_history(iteration_ids, dates)
}
\arguments{
\item{iteration_ids}{A comma-separated list of iteration IDs. Use \code{\link{get_iteration_tree}} to acquire a list of iterations.}

\item{dates}{A list of \code{date} objects representing the points in time to calculate the backlog size.}
}
\value{
Dataframe with three columns: \code{TOTAL_RELEASE_POINTS}, \code{TOTAL_RELEASE_COUNT} and \code{AS_OF} (indicating the date)
}
\description{
Get the total backlog size at different points in time. Only considers work items under the team's default area path.
}
\examples{
iteration_ids <- '50, 51, 52'
dates <- list(Sys.Date(), Sys.Date() - 14, Sys.Date() - 28)
get_backlog_history(iteration_ids, dates)
}

