% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_current_release}
\alias{get_current_release}
\title{Get Current Release}
\usage{
get_current_release()
}
\value{
A list of current major (e.g., 3.0) and minor (e.g., 3.1) releases.
Current minor release will be \code{NULL} if there are no minor releases defined.
}
\description{
Computes the current release based on today's date.
}

