% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_default_area_path}
\alias{get_default_area_path}
\title{Get Default Area Path}
\usage{
get_default_area_path()
}
\value{
Default area path string
}
\description{
Get TFS Team's default area path.
}
\examples{
get_default_area_path()
}

