% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_iteration_tree}
\alias{get_iteration_tree}
\title{Gets iteration tree for a particular TFS project}
\usage{
get_iteration_tree()
}
\arguments{
\item{None.}{}
}
\value{
\code{api_get} class of iterations
}
\description{
Gets iteration tree for a particular TFS project
}

