% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_planned_velocity}
\alias{get_planned_velocity}
\title{Get Planned Velocity}
\usage{
get_planned_velocity(iteration_id, date_time)
}
\arguments{
\item{iteration_id}{The iteration ID of the sprint whose planned velocity is desired.}

\item{date_time}{The date and time to calculate the planned velocity.}
}
\value{
Dataframe with two columns: \code{PLANNED_VELOCITY} and \code{SPRINT_ITERATION_ID}
}
\description{
Get planned velocity for each sprint. Only considers work items under the team's default area path.
}
\examples{
iteration_id <- '50'
date_time <- update(as_datetime('2016-09-01T00:00:00Z'), hour = 23, minute = 59)
get_planned_velocity(iteration_id, date_time)
}

