% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_projects}
\alias{get_projects}
\title{Get TFS Projects}
\usage{
get_projects(tfs_collection)
}
\arguments{
\item{tfs_collection}{TFS Collection}
}
\value{
\code{api_get} class of TFS Projects
}
\description{
Get a list of projects in a TFS Collection.
}
\examples{
get_projects('FinancialReportingCollection')
}

