% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_summary.R
\name{get_release_feature_completion}
\alias{get_release_feature_completion}
\title{Get Release Feature Completion}
\usage{
get_release_feature_completion(release_iteration_id, dates)
}
\value{
Dataframe with 6 columns: \code{ID}, \code{TITLE}, \code{PERCENT_COMPLETE}, \code{TOTAL_POINTS}, \code{COMPLETED_POINTS} and \code{AS_OF}
}
\description{
Get completion percentages for each feature over the course of a release.
Percentage is based on sum of child work items completed effort divided by total effort.
}

