% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_release_feature_ids}
\alias{get_release_feature_ids}
\title{Get Release Feature IDs}
\usage{
get_release_feature_ids(iteration_id)
}
\value{
Dataframe with 3 columns: \code{ID}, \code{FEATURE TITLE} and \code{PERCENT COMPLETE}
}
\description{
Get completion percentages for each feature marked with a particular iteration path / release.
Percentage is based on sum of child work items completed effort divided by total effort.
}

