% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.R
\name{get_release_wis}
\alias{get_release_wis}
\title{Get Release Work Items}
\usage{
get_release_wis(wi_id_list, date = format(Sys.Date()))
}
\arguments{
\item{wi_id_list}{A comma-separated list of work item IDs. Use \code{\link{get_release_wi_ids}} to acquire a list of work items.}

\item{date}{The TFS API has the ability to fetch data as of a certain date. Passing a date object will return a list of work items
as of this date. Defaults to today's date.}
}
\value{
\code{api_post} class of work items
}
\description{
Get a list of work items for a particular team's release.
}
\examples{
wi_id_list <- '50000, 51000, 52000'
get_release_wis(wi_id_list)
}

