% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprint_summary.R
\name{get_sprint_history}
\alias{get_sprint_history}
\title{Get Sprint History}
\usage{
get_sprint_history(iteration_id, sprint_start_date, sprint_end_date)
}
\arguments{
\item{iteration_id}{An iteration ID. Use \code{\link{get_iteration_tree}} to acquire a list of iterations.}

\item{dates}{A list of \code{date} objects representing the points in time to calculate the backlog size.}
}
\value{
Dataframe with five columns: \code{COMPLETED_POINTS}, \code{TOTAL_POINTS}, \code{COMPLETED_COUNT}, \code{TOTAL_COUNT} and \code{AS_OF} (indicating the date)
}
\description{
Get the cumulative completed work item points on each day of the current sprint.
}
\examples{
iteration_id <- '50'
sprint_start_date <- '2016-01-01'
sprint_end_date <- '2016-01-14'
get_sprint_history(iteration_id, sprint_start_date, sprint_end_date)
}

