% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases.R
\name{get_tc_automation_history}
\alias{get_tc_automation_history}
\title{Get Test Case Automation History}
\usage{
get_tc_automation_history(dates)
}
\arguments{
\item{dates}{A list of \code{date} objects representing the points in time to calculate the backlog size.}
}
\value{
Dataframe with three columns: \code{TEST_CASES_TOTAL}, \code{TEST_CASES_AUTOMATED} and \code{AS_OF} (indicating the date)
}
\description{
Gets the total number of test cases and the number of test cases that are automated on different dates.
Only considers test cases under the team's default area path.
}
\examples{
dates <- list(Sys.Date(), Sys.Date() - 14, Sys.Date() - 28)
get_tc_automation_history(dates)
}

