% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_team_area_paths}
\alias{get_team_area_paths}
\title{Get Team Area Path(s)}
\usage{
get_team_area_paths()
}
\value{
List of team area paths
}
\description{
Get TFS Team's assigned area path(s), including the default area path.
}
\examples{
get_team_area_paths()
}

