% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_team_backlog_filter}
\alias{get_team_backlog_filter}
\title{Get Team Backlog Filter}
\usage{
get_team_backlog_filter()
}
\value{
Response from call to TFS TeamFieldValues.
}
\description{
Gets a TFS Team's backlog filter type (TR.Elite.Team or System.AreaPath) and values.
}
\examples{
get_team_backlog_filter()
}

