% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_team_backlog_filter_wiql}
\alias{get_team_backlog_filter_wiql}
\title{Get Team Backlog Filter WIQL}
\usage{
get_team_backlog_filter_wiql()
}
\value{
WIQL string to be included in WIQL calls to REST API.
}
\description{
Gets the WIQL of the team's backlog filter.
}
\examples{
get_team_backlog_filter_wiql()
}

