% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_projects_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get TFS Project teams}
\usage{
get_teams(tfs_collection, tfs_project)
}
\arguments{
\item{tfs_collection}{TFS Collection}

\item{tfs_project}{TFS Project that is in \code{tfs_collection}. Use \code{\link{get_projects}} for a list of projects in a collection.}
}
\value{
\code{api_get} class of TFS Teams
}
\description{
Get a list of teams in a TFS Project.
}
\examples{
get_teams('FinancialReportingCollection', 'FinancialReportingProjects')
}

