% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases.R
\name{get_test_case_ids}
\alias{get_test_case_ids}
\title{Get Test Case IDs}
\usage{
get_test_case_ids(date = format(Sys.Date()))
}
\arguments{
\item{date}{The TFS API has the ability to fetch data as of a certain date. Passing a date object will return a
list of test case IDs whose area path is under \code{get_default_area_path} as of this date. Defaults to today's date.}
}
\value{
\code{api_post} class of test case IDs
}
\description{
Get a list of test case IDs for a particular product.
}
\examples{
get_test_case_ids()
}

