% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases.R
\name{get_test_cases}
\alias{get_test_cases}
\title{Get Test Cases}
\usage{
get_test_cases(test_case_id_list, date = format(Sys.Date()))
}
\arguments{
\item{test_case_id_list}{A comma-separated list of test case IDs. Use \code{\link{get_test_case_ids}} to acquire a list of test cases.}

\item{date}{The TFS API has the ability to fetch data as of a certain date. Passing a date object will return a list of work items
as of this date. Defaults to today's date.}
}
\value{
\code{api_post} class of test cases
}
\description{
Get a list of test cases for a particular product. Currently uses the team's default area path.
}
\examples{
test_case_id_list <- '50000, 51000, 52000'
get_test_cases(wi_id_list)
}

